+++ examples/complete/main.tf
@@ -1,6 +1,7 @@
 module "aci_monitoring_policy" {
-  source = "netascode/monitoring-policy/aci"
+  source  = "netascode/monitoring-policy/aci"
+  version = ">= 0.0.2"
 
-  snmp_policy_name   = "SNMP1"
-  syslog_policy_name = "SYSLOG1"
+  snmp_trap_policy = "SNMP1"
+  syslog_policy    = "SYSLOG1"
 }