+++ variables.tf
@@ -1,21 +1,21 @@
-variable "snmp_policy_name" {
-  description = "SNMP policy name."
+variable "snmp_trap_policy" {
+  description = "SNMP trap policy name."
   type        = string
   default     = ""
 
   validation {
-    condition     = can(regex("^[a-zA-Z0-9_.-]{0,64}$", var.snmp_policy_name))
+    condition     = can(regex("^[a-zA-Z0-9_.-]{0,64}$", var.snmp_trap_policy))
     error_message = "Allowed characters: `a`-`z`, `A`-`Z`, `0`-`9`, `_`, `.`, `-`. Maximum characters: 64."
   }
 }
 
-variable "syslog_policy_name" {
+variable "syslog_policy" {
   description = "Syslog policy name."
   type        = string
   default     = ""
 
   validation {
-    condition     = can(regex("^[a-zA-Z0-9_.-]{0,64}$", var.syslog_policy_name))
+    condition     = can(regex("^[a-zA-Z0-9_.-]{0,64}$", var.syslog_policy))
     error_message = "Allowed characters: `a`-`z`, `A`-`Z`, `0`-`9`, `_`, `.`, `-`. Maximum characters: 64."
   }
 }