+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+- BREAKING CHANGE: rename snmp_policy_name variable to snmp_trap_policy
+- BREAKING CHANGE: rename syslog_policy_name variable to syslog_policy
+
 ## 0.0.1
 
 - Initial release